@@ -1,7 +1,8 @@
 import "./app.css";
+import { mount } from "svelte";
 import App from "./App.svelte";
 
-const app = new App( {
+const app = mount( App, {
 	target: document.querySelector( "body" ) as Element
 } );
 
